+++ src/app/app.component.ts
@@ -19,7 +19,7 @@ import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 
 @Component({
-  selector: 'bed-occupancy-root',
+  selector: 'app-bed-occupancy-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
   imports: [RouterOutlet],