+++ src/app/bed-occupancy-new/bed-occupancy-question/bed-occupancy-question.component.ts
@@ -15,27 +15,45 @@
     find details in the "Readme" file.
  */
 
+import { NgTemplateOutlet } from '@angular/common';
 import { Component, inject } from '@angular/core';
-import { MaxHeightContentContainerComponent, SectionHeaderComponent, StepContentComponent, StepNavigation } from '@gematik/demis-portal-core-library';
-import { FormlyForm } from '@ngx-formly/core';
-import { BedOccupancyNotificationService } from '../bed-occupancy-notification.service';
+import { ReactiveFormsModule } from '@angular/forms';
 import { MatButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { MatToolbar, MatToolbarRow } from '@angular/material/toolbar';
+import {
+  GemDemisAriaDisabledButtonDirective,
+  MaxHeightContentContainerComponent,
+  SectionHeaderComponent,
+  StepContentComponent,
+  StepNavigation,
+} from '@gematik/demis-portal-core-library';
+import { FormlyForm } from '@ngx-formly/core';
+import { environment } from 'src/environments/environment';
 import { BedOccupancyConstants } from '../../bed-occupancy/common/bed-occupancy-constants';
 import { questionBedOccupancyHtmlConfigFieldsNew } from '../../shared/formly/configs/bed-occupancy/question-new.config';
-import { environment } from 'src/environments/environment';
-import { NgTemplateOutlet } from '@angular/common';
+import { BedOccupancyNotificationService } from '../bed-occupancy-notification.service';
 
 @Component({
   selector: 'app-bed-occupancy-question',
-  imports: [FormlyForm, MatButton, MatIcon, MatToolbar, MatToolbarRow, SectionHeaderComponent, MaxHeightContentContainerComponent, NgTemplateOutlet],
+  imports: [
+    FormlyForm,
+    MatButton,
+    MatIcon,
+    MatToolbar,
+    MatToolbarRow,
+    SectionHeaderComponent,
+    MaxHeightContentContainerComponent,
+    NgTemplateOutlet,
+    GemDemisAriaDisabledButtonDirective,
+    ReactiveFormsModule,
+  ],
   templateUrl: './bed-occupancy-question.component.html',
   styleUrl: './bed-occupancy-question.component.scss',
 })
 export class BedOccupancyQuestionComponent extends StepContentComponent<void> {
   protected readonly notificationService = inject(BedOccupancyNotificationService);
-  protected readonly fieldConfig = questionBedOccupancyHtmlConfigFieldsNew;
+  protected readonly fieldConfig = questionBedOccupancyHtmlConfigFieldsNew();
   protected navigation = inject(StepNavigation);
   protected readonly BedOccupancyConstants = BedOccupancyConstants;