+++ src/single-spa/asset-url.ts
@@ -19,8 +19,10 @@
 // instead of hard coded to `/assets`. We use webpack public path for this.
 // See https://webpack.js.org/guides/public-path/#root
 
+// Injected by webpack at runtime; see https://webpack.js.org/guides/public-path/#root
+declare const __webpack_public_path__: string;
+
 export function assetUrl(url: string): string {
-  // @ts-ignore
   const publicPath = __webpack_public_path__;
   const publicPathSuffix = publicPath.endsWith('/') ? '' : '/';
   const urlPrefix = url.startsWith('/') ? '' : '/';