+++ src/test/integration/bed-occupancy-new-sidenav.integration.spec.ts
@@ -105,7 +105,7 @@ describe.skip('BedOccupancy with new sidenav Integration and Playwright', () => 
     expect(submitButton.disabled, `Submit button should be disabled for value: ${testValue}`).toBe(true);
   }
 
-  async function checkDescribingError(_input: MatInputHarness, expectedResult: String) {
+  async function checkDescribingError(_input: MatInputHarness, expectedResult: string) {
     await fixture.whenStable();
     fixture.detectChanges();
     const formlyErrors = Array.from(fixture.nativeElement.querySelectorAll('mat-error formly-validation-message')) as HTMLElement[];